+++ packages/components/src/ui/index.js
@@ -3,7 +3,6 @@ export * from './control-group';
 export * from './control-label';
 export * from './divider';
 export * from './elevation';
-export * from './flex';
 export * from './form-group';
 export * from './grid';
 export * from './h-stack';